Move TextureReplacer to Core.
Probably makes more sense here after all.
